+++ src/Symfony/Component/Security/Http/LoginLink/LoginLinkHandlerInterface.php
@@ -24,12 +24,16 @@ interface LoginLinkHandlerInterface
     /**
      * Generate a link that can be used to authenticate as the given user.
      *
-     * @param int|null $lifetime When not null, the argument overrides any default lifetime previously set
+     * @param int|null                          $lifetime   When not null, the argument overrides any default lifetime previously set
+     * @param array<string, \Stringable|scalar> $parameters A list of additional query string parameters that should be part of the login link URL
      */
-    public function createLoginLink(UserInterface $user, ?Request $request = null, ?int $lifetime = null): LoginLinkDetails;
+    public function createLoginLink(UserInterface $user, ?Request $request = null, ?int $lifetime = null/* , array $parameters = [] */): LoginLinkDetails;
 
     /**
      * Validates if this request contains a login link and returns the associated User.
+     *
+     * Implementations should expose the extra parameters covered by the link signature
+     * via the "_login_link_parameters" request attribute.
      *
      * Throw InvalidLoginLinkExceptionInterface if the link is invalid.
      */