+++ src/Symfony/Component/Security/Http/CHANGELOG.md
@@ -14,6 +14,8 @@ CHANGELOG
  * Add `ImpersonateUrlGenerator::generateImpersonationForm()` and `generateExitForm()` to build a form that switches the user with a POST
  * Add `$targetUri` argument to `ImpersonateUrlGenerator::generateImpersonationPath()` and `generateImpersonationUrl()`
  * Configure the decorated handler of `CustomAuthenticationSuccessHandler` and `CustomAuthenticationFailureHandler` when they are called instead of when they are built, so that a single handler can be shared by several authenticators
+ * Add argument `$parameters` to `LoginLinkHandlerInterface::createLoginLink()` to add extra query parameters covered by the link signature
+ * Expose the verified extra parameters via the `_login_link_parameters` request attribute when consuming a login link
 
 8.1
 ---